pyproject.toml
[tool.poetry]
name = "disciplined"
version = "0.1.0"
description = ""
authors = ["Nishant Kumar"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
